refactor pages and simplify gulp tasks, removing need for pages
